Renames package countries to occupations
