+++ core/src/main/occupations/dealers/CountryDealerExceptions.kt
@@ -1,4 +1,4 @@
-package countries.dealers
+package occupations.dealers
 
 class NoPlayersToDealToException : Exception(
     "Can't deal to an empty collection of players.")
